requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websocket-client==1.6.4
pydantic==2.5.0
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.4
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1